requirements.txt
opencv-python>=4.8.0
numpy>=1.23.0
scikit-learn>=1.3.0
PyYAML>=6.0
pyinstaller>=6.0
